router: moved all the metrics code to metrics.go
In passing, apply reviewer's suggestion of replacing the init()
function with a static assertion.
